+++ src/CQChartsSankeyPlot.cpp
@@ -18,10 +18,10 @@ addParameters()
 {
   startParameterGroup("Sankey");
 
-  addColumnParameter("link", "Source/Target", "linkColumn" , 0).
+  addColumnParameter("link", "Source/Target", "linkColumn").
     setRequired().setTip("Name pair for Source/Target connection");
 
-  addColumnParameter("value", "Value", "valueColumn", 1).
+  addColumnParameter("value", "Value", "valueColumn").
     setRequired().setTip("Connection value");
 
   endParameterGroup();